Add support for RDEId in the CAPI Yaml specification (#63)
Problem: Currently, applying CAPI yaml on the management cluster is the only entry point for the creation of workload cluster. In order to enable RDE creation as an entry point in the future, we need to let CAPVCD use the pre-created RDEId for any cluster request to be processed.

Jira: VCDA-3381

Solution: Add support for RDEId to come as input in the CAPI Yaml. CAPVCD will use it if it is non-empty (or) it will create a new RDE-ID

Signed-off-by: sayloo <sahithi.ayloo@gmail.com>