VCDA-3722,3828,3835: Use user-specified IP, remove oneArm usage, (#170)
ensure that right params are passed for LB creation
Fixes #149

To use a custom control-plane IP and port, the following parameters need to be set in the vcdClusterObject as follows:

vcdCluster:
  spec:
    controlPlaneEndpoint:
      host: a.b.c.d
      port: 6443
Any routable IP can be used for the host.

Also add a `useDNAT` parameter in the `loadBalancer` section to indicate that DNAT+OneArm need to be used. It is disabled by default in v1beta1 and enabled by default in v1alpha4.