Allow users to define vm naming logic (#369)
* Allow users to define vm naming logic

Now, VMs are named as machine.Name. In some cases,
we need to name VMs with a custom logic.

This PR adds a new field into VCDMachine CR so that
users can define templates to name VMs by using Go templates
and Sprig functions. ".machine" and ".vcdMachine" will refer
to Machine and VCDMachine CRs in the templates.

* Fix template function map

* Remove duplicate SetupWebhookWithManager call

* Remove unnecessary nil check

* Fix omit empty issues