Image updates to reduce image size, move to golang static binary and scratch image (#446)
* chore: fix image layers

fix image layers to prevent unneeded layer. By moving the `chmod +x` to before the copy, we don't have to include that layer in the changes
```bash
test2                                                                                     latest            32186ae3133f   35 minutes ago   93.8MB
test                                                                                      latest            af43cf4b9444   35 minutes ago   151MB
```

If we inspect the layers we can see the copy layer is 57MB and the chmod layer is also 57MB.
```bash
57 MB  COPY /build/vcloud/cluster-api-provider-cloud-director . # buildkit
57 MB  RUN /bin/sh -c chmod +x /opt/vcloud/bin/cluster-api-provider-cloud-director # buildkit
```

after the change we don't have the 2nd layer

```bash
57 MB  COPY /build/vcloud/cluster-api-provider-cloud-director . # buildkit
```

Signed-off-by: slimm609 <dbrian@vmware.com>

* fix: remove debugging symbols and compile static

- remove the debugging symbols and gdb information from binary and change to a static binary.
This reduces the size by 17MB

```bash
57 MB  COPY /build/vcloud/cluster-api-provider-cloud-director . # buildkit
```

```bash
40 MB  COPY /build/vcloud/cluster-api-provider-cloud-director . # buildkit

```

Signed-off-by: slimm609 <dbrian@vmware.com>

* fix: move to a scratch image

update to a scratch image with only cacerts. Since its a static binary, the photon base is not needed.

---------

Signed-off-by: slimm609 <dbrian@vmware.com>